Added game framework and image blurring
The GameActivity class sets the Game class (which extends Surface View)
as its content view. The Game renders a bitmap image which slowly gets
more and more blurry.

The BlurFilter class handles the blurring. It takes each pixel's
neighboring pixels color values and calculates an average. This average
color is what the pixel updates its value to. When running the program
in the emulator, the blurring is extremly slow.