Added buttons and image swapping
The letters are clickable and will then be placed on the leftmost empty
"word letter box". The letters can also be dragged to any of the empty
word letter boxes.

When a word letter box is clicked, its letter moves back to the letter
boxes at the bottom of the display.

A program which blurs and saves images has been made. It will be further
developed later. The 23 different images (generated from the program)
are getting swapped depending on how long time has passed. This is yet
not complete. The swapping of the images needs to get smoother.

The source code of the image blurrer will be added into this repository
soon.