Game framework done
The concept of the game is now done. Images are being swapped depending
on how long the game has been running. To get an animation between the
image swaps, an image with lower opacity is being drawn over the current
image. The opacity of the covering image depends on how much time it is
left until the image is being swapped.

For testing purposes, the background color is being changed to black
when the time is up and gray when the correct word is typed.

It is also possible to use spaces in the answers. The spaces will not
generate a word button.