Added onPause, onResume etc.
When the game is closed (stopped) the Home Activity is being started
once the application is being started again. There are still a few bugs
which will be fixed.

Somewhere in the code, there is a memory leak which makes the phone to
run out of energy. To slow this process down, all the bitmaps used in an
activity are set to null before starting a new activity.

Furthermore, images in the Image Blurring project are now being cropped.
The ratio is 5 : 7.