Added games and a score system
The score system is very badly made and probably extremely easy to hack.
It is just a test and will be developed further. When a new game is
made, the opponent is always "Po". When a game is played the time it
took for the player to guess the image is multiplied by 100 and sent to
the gamefinished.php. There the score is inserted into the correct row
in the MySQL database. The last_played_time is also updated to the
current timestamp.

In the StartGameActivity, all the games the user plays are loaded via
the getgames.php from the MySQL database. When a game is clicked, the
GameOverviewActivity shows a overview of the game. There a play button
can be pressed with starts a game.