Improved game and score system
The games now set images that will be used by both the users. This data
is sent unprotected and will probably need to be encrypted in the future
to prevent cheaters from knowing what the correct words are.

The score of each game is determined by the formula
s = 1000 - 1000 * t / 45
where s is the score and t is the time in seconds. If the correct word
has not been guessed after 45 seconds, the user has lost.