Reworked the game score system
The player now plays five games in a row and gets a total score. This
score is sent to the database and the game_state field is changed so
that the other player can play. When a round ends, the player with the
most points gets +1 score. This score is what is being displayed in the
Start Game Activity.

If the player would leave the application while in a game, the current
score is sent to the database. When re-entering the application, the
player is redirected to the Start Game Activity.