Moved inner classes out
The UpdateGames class has been moved out to a separate .java-file. It
can now be accessed by other classes as well.

The getimage.php is now used by the startgame.php and gamefinished.php.
The code has moved from the startgame.php back to the getimage.php. This
was done to make it accessible by all the .php-files.

Before a game is started, it is checked against the database if it still
is the player's turn. If not, the information is updated instead.