Improved security
The improvement is very small. When a player checks for the game state,
it is updated in the database at which time this occurs. The score then
has six minutes to be sent back. If the time has been exeeded, the score
is set to 0. This will probably only occur if a player tries to cheat.

Also, the player is alerted when it is not possible to update the games
or start one due to connection issues.